[CBR-150] Implement rollback
Also wrote a hand-written test to test specifically for dependent pending
transactions, because the generator doesn't always hit that case (and fixed the
bug that Ryan intentionally left in because he rightly wanted a test first that
exposes it first).

Also fixed a bug in the test infrastructure: the interpreter requires state
that depends on the blockchain (hash of the previous block, the stake
distrubtion, etc). It is therefore important that we correctly track this state
and rollback the interpreter state when there is a rollback in the blockchain.

Added shrinking to test comparing the kernel to the pure model.